requirements.txt
streamlit>=1.32.0
streamlit-authenticator>=0.3.2
streamlit-option-menu>=0.3.12
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
pyyaml>=6.0
bcrypt>=4.1.0
python-dateutil>=2.8.0
